requirements.txt
lgpio==0.2.2.0
python-dotenv==1.1.1
rpi-lgpio==0.6
